feat: simplify message broadcast channel
