+++ core/src/main/java/com/trionesdev/phecda/foundation/core/domains/messageforwarding/dao/impl/MessageSinkDAO.java
@@ -4,7 +4,7 @@ import cn.hutool.core.collection.CollectionUtil;
 import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
 import com.trionesdev.phecda.foundation.core.domains.messageforwarding.dao.po.MessageSinkPO;
-import com.trionesdev.phecda.foundation.core.domains.messageforwarding.internal.enums.SinkActionType;
+import com.trionesdev.phecda.foundation.core.domains.messageforwarding.shared.enums.SinkActionType;
 import com.trionesdev.phecda.foundation.core.domains.messageforwarding.dao.mapper.MessageSinkMapper;
 
 import java.util.Collection;
