+++ core/src/main/java/com/trionesdev/phecda/foundation/core/domains/messageforwarding/internal/MessageForwardingBootstrap.java
@@ -1,18 +1,12 @@
 package com.trionesdev.phecda.foundation.core.domains.messageforwarding.internal;
 
-import com.trionesdev.message.core.Message;
 import com.trionesdev.message.core.MessageContainer;
-import com.trionesdev.message.core.MessageListener;
 import com.trionesdev.phecda.foundation.core.domains.messageforwarding.service.factory.ForwardingActionFactory;
+import com.trionesdev.phecda.foundation.core.domains.messageforwarding.service.impl.MessageSinkService;
 import lombok.RequiredArgsConstructor;
-import com.trionesdev.phecda.foundation.core.domains.messageforwarding.internal.event.redis.RuleChangeListener;
-import com.trionesdev.phecda.foundation.core.domains.messageforwarding.internal.event.redis.SinkActionChangeListener;
 import org.springframework.boot.ApplicationArguments;
 import org.springframework.boot.ApplicationRunner;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.data.redis.listener.PatternTopic;
-import org.springframework.data.redis.listener.RedisMessageListenerContainer;
-import org.springframework.data.redis.listener.adapter.MessageListenerAdapter;
 
 import java.util.concurrent.CompletableFuture;
 
@@ -22,31 +16,19 @@ import static com.trionesdev.phecda.foundation.core.domains.messageforwarding.in
 @RequiredArgsConstructor
 @Configuration
 public class MessageForwardingBootstrap implements ApplicationRunner {
-    private final RedisMessageListenerContainer redisMessageListenerContainer;
-    private final RuleChangeListener ruleChangeListener;
-    private final SinkActionChangeListener sinkActionChangeListener;
     private final MessageContainer messageContainer;
     private final ForwardingActionFactory forwardingActionFactory;
+    private final MessageSinkService messageSinkService;
 
     @Override
     public void run(ApplicationArguments args) {
         CompletableFuture.runAsync(() -> {
-            redisMessageListenerContainer.addMessageListener(new MessageListenerAdapter(ruleChangeListener, "onMessage"), PatternTopic.of(MESSAGE_FORWARDING_RULE_CHANGE));
-            redisMessageListenerContainer.addMessageListener(new MessageListenerAdapter(sinkActionChangeListener, "onMessage"), PatternTopic.of(MESSAGE_SINK_CHANGE));
+            messageContainer.addBroadcastListener(message -> forwardingActionFactory.syncMessageForwardingRules(), MESSAGE_FORWARDING_RULE_CHANGE);
 
-//            messageContainer.addBroadcastListener(new MessageListener() {
-//                @Override
-//                public void onMessage(Message message) {
-//                    forwardingActionFactory.syncMessageForwardingRules();
-//                }
-//            }, MESSAGE_FORWARDING_RULE_CHANGE);
-//
-//            messageContainer.addBroadcastListener(new MessageListener() {
-//                @Override
-//                public void onMessage(Message message) {
-//                    String id = message.getPayload();
-//                }
-//            }, MESSAGE_SINK_CHANGE);
+            messageContainer.addBroadcastListener(message -> {
+                String sinkId = message.getPayload();
+                messageSinkService.syncSinkActions(sinkId);
+            }, MESSAGE_SINK_CHANGE);
         });
     }
 }
