+++ core/src/main/java/com/trionesdev/phecda/foundation/core/domains/messageforwarding/internal/aggregate/entity/MessageSource.java
@@ -1,6 +1,6 @@
 package com.trionesdev.phecda.foundation.core.domains.messageforwarding.internal.aggregate.entity;
 
-import com.trionesdev.phecda.foundation.core.domains.messageforwarding.internal.model.source.SourceProps;
+import com.trionesdev.phecda.foundation.core.domains.messageforwarding.shared.model.source.SourceProps;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
