--- core/src/main/java/com/trionesdev/phecda/foundation/core/domains/messageforwarding/internal/enums/SinkActionType.java
@@ -1,5 +1,0 @@
-package com.trionesdev.phecda.foundation.core.domains.messageforwarding.internal.enums;
-
-public enum SinkActionType {
-    KAFKA
-}
