--- core/src/main/java/com/trionesdev/phecda/foundation/core/domains/messageforwarding/internal/model/source/MessageType.java
@@ -1,6 +1,0 @@
-package com.trionesdev.phecda.foundation.core.domains.messageforwarding.internal.model.source;
-
-public enum MessageType {
-    THING_PROPERTY_REPORT,
-    THING_EVENT_REPORT
-}
