+++ core/src/main/java/com/trionesdev/phecda/foundation/core/domains/messageforwarding/service/impl/MessageSinkService.java
@@ -2,28 +2,32 @@ package com.trionesdev.phecda.foundation.core.domains.messageforwarding.service.
 
 import cn.hutool.core.collection.CollectionUtil;
 import com.trionesdev.commons.exception.BusinessException;
+import com.trionesdev.message.core.MessageContainer;
 import com.trionesdev.phecda.foundation.core.domains.messageforwarding.dao.po.MessageSinkPO;
-import lombok.RequiredArgsConstructor;
 import com.trionesdev.phecda.foundation.core.domains.messageforwarding.dao.po.RuleSinkPO;
-import com.trionesdev.phecda.foundation.core.domains.messageforwarding.internal.event.spring.MessageSinkChangeEvent;
 import com.trionesdev.phecda.foundation.core.domains.messageforwarding.manager.impl.MessageSinkManager;
 import com.trionesdev.phecda.foundation.core.domains.messageforwarding.manager.impl.RuleSinkManager;
-import org.springframework.context.ApplicationEventPublisher;
+import com.trionesdev.phecda.foundation.core.domains.messageforwarding.service.factory.action.KafkaForwardingAction;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
+
+import static com.trionesdev.phecda.foundation.core.domains.messageforwarding.internal.event.EventConstants.MESSAGE_SINK_CHANGE;
 
 @RequiredArgsConstructor
 @Service
 public class MessageSinkService {
-    private final ApplicationEventPublisher applicationEventPublisher;
+    private final MessageContainer messageContainer;
     private final MessageSinkManager messageSinkManager;
     private final RuleSinkManager ruleSinkManager;
+    private final KafkaForwardingAction kafkaForwardingAction;
 
     public void create(MessageSinkPO record) {
         messageSinkManager.create(record);
-        applicationEventPublisher.publishEvent(new MessageSinkChangeEvent(this, record));
+        messageContainer.broadcast(MESSAGE_SINK_CHANGE, record.getId());
     }
 
     public void deleteById(String id) {
@@ -33,14 +37,14 @@ public class MessageSinkService {
         }
         messageSinkManager.findById(id).ifPresent(messageSinkSnap -> {
             messageSinkManager.deleteById(id);
-            applicationEventPublisher.publishEvent(new MessageSinkChangeEvent(this, messageSinkSnap));
+            messageContainer.broadcast(MESSAGE_SINK_CHANGE, messageSinkSnap.getId());
         });
 
     }
 
     public void updateById(MessageSinkPO record) {
         messageSinkManager.updateById(record);
-        applicationEventPublisher.publishEvent(new MessageSinkChangeEvent(this, record));
+        messageContainer.broadcast(MESSAGE_SINK_CHANGE, record.getId());
     }
 
     public Optional<MessageSinkPO> findById(String id) {
@@ -49,6 +53,19 @@ public class MessageSinkService {
 
     public List<MessageSinkPO> findList() {
         return messageSinkManager.findList();
+    }
+
+    public void syncSinkActions(String id) {
+        messageSinkManager.findById(id).ifPresent(messageSink -> {
+            if (Objects.isNull(messageSink.getAction())) {
+                return;
+            }
+            switch (messageSink.getAction().getType()) {
+                case KAFKA:
+                    kafkaForwardingAction.kafkaSync();
+                    break;
+            }
+        });
     }
 
 }
