+++ core/src/main/java/com/trionesdev/phecda/foundation/core/domains/messageforwarding/shared/enums/SinkActionType.java
@@ -1,0 +1,5 @@
+package com.trionesdev.phecda.foundation.core.domains.messageforwarding.shared.enums;
+
+public enum SinkActionType {
+    KAFKA
+}
