+++ core/src/main/java/com/trionesdev/phecda/foundation/core/domains/messageforwarding/shared/model/source/MessageType.java
@@ -1,0 +1,6 @@
+package com.trionesdev.phecda.foundation.core.domains.messageforwarding.shared.model.source;
+
+public enum MessageType {
+    THING_PROPERTY_REPORT,
+    THING_EVENT_REPORT
+}
