+++ core/src/main/java/com/trionesdev/phecda/foundation/core/domains/messageforwarding/shared/model/source/SourceProps.java
@@ -1,4 +1,4 @@
-package com.trionesdev.phecda.foundation.core.domains.messageforwarding.internal.model.source;
+package com.trionesdev.phecda.foundation.core.domains.messageforwarding.shared.model.source;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonSubTypes;
