+++ interfaces/rest-tenant/src/main/java/com/trionesdev/phecda/foundation/rest/tenant/domains/messageforwarding/controller/ro/MessageSinkUpdateRO.java
@@ -1,8 +1,8 @@
 package com.trionesdev.phecda.foundation.rest.tenant.domains.messageforwarding.controller.ro;
 
-import com.trionesdev.phecda.foundation.core.domains.messageforwarding.internal.enums.SinkActionType;
+import com.trionesdev.phecda.foundation.core.domains.messageforwarding.shared.enums.SinkActionType;
 import lombok.Data;
-import com.trionesdev.phecda.foundation.core.domains.messageforwarding.internal.model.sinkaction.SinkActionProps;
+import com.trionesdev.phecda.foundation.core.domains.messageforwarding.shared.model.sinkaction.SinkActionProps;
 
 @Data
 public class MessageSinkUpdateRO {
