+++ interfaces/rest-tenant/src/main/java/com/trionesdev/phecda/foundation/rest/tenant/domains/messageforwarding/controller/ro/MessageSourceTopicRO.java
@@ -1,7 +1,7 @@
 package com.trionesdev.phecda.foundation.rest.tenant.domains.messageforwarding.controller.ro;
 
 import lombok.Data;
-import com.trionesdev.phecda.foundation.core.domains.messageforwarding.internal.model.source.SourceProps;
+import com.trionesdev.phecda.foundation.core.domains.messageforwarding.shared.model.source.SourceProps;
 
 public class MessageSourceTopicRO {
     @Data
